move functions into tidier files
